feat: Controller 에서 체스판 모양을 만들어주는 기능을 Dto로 위임
